Start ripping things apart
