+++ wizwalker/memory/__init__.py
@@ -1,6 +1,5 @@
 from .handler import HookHandler
 from .hooks import *
 from .memory_object import MemoryObject
-from .memory_reader import MemoryReader
 from .memory_objects import *
 from .instance_finder import InstanceFinder
